Make install-results cope with old build-directories.
When install-results moves the final output directories from
.../.tmp/X to .../X, it fails if there's already a directory with the
given name in its target directory; this arises when repeating a
build.  So remove any such conflicting directories from the
destination, before trying to move their replacements in.

At the same time, avoid repetition of long and tedious paths by using
variables (which makes it easier to see which ones are the same and
makes the differences easier to see) and skip spurious braces.  Put
all variable expansion inside quotes and ensure that names with odd
things (like spaces) in them won't cause problems.

Given that mkdir -p creates parents, one of the invocations of it
could be skipped, as it made the parent of the next.