Disable checksum for LMDB tarball.
Since it is downloaded directly from the git tarball generator, its
checksum may change due to differing timestamps or internal file
order. But we are downloading a direct Git SHA, so we trust that the
contents are the same.

(cherry picked from commit 5765bdeef0bd0c74b33340a480047afe933816be)