Redmine #6581: Don't perform uninstall steps if we are upgrading.
This fixes the daemons being stopped after an upgrade.

Note that this will only have an effect when we upgrade from a package
that includes this change, because the uninstall script is always
taken from the *old* package. Hence, the earliest package that this
change will have any effect for, is the second version that includes
the change, which at the time of writing is probably going to be
3.6.3.

(cherry picked from commit 8dd4e122c107f59b014055b09ed0b0b5e43553ed)