Repeat fix 72ed4e0a for all dependencies.
As it turns out all Solaris dependencies suffer from the same problem,
so fix them all (and do FreeBSD as well, while we're at it). The fix
is:

Make sure all .so files are part of run time dependencies. The old
code omitted .so without a numbered suffix, which would make it a
build time dependency instead of run time dependency.