Revert "add all dependencies needed by libgcc and associates for aix"
This reverts commit 19721c335ae2b8cc6edadeefb1d3eb722c8b60db.

It introduces way more problems than it solves, and in general it is
not safe to assume that the very platform specific libraries libc and
libpthread are going to work on all platforms we install the AIX
package on. This unfortunately also means that we cannot support AIX
platforms with a lower patch level than our build machine.

Redmine #6679.

Conflicts:

	deps-packaging/libgcc/cfbuild-libgcc.spec