AIX: Make sure cf-agent also gets killed when calling "cfengine3 stop"
cf-execd can call cf-agent right before it gets killed, and if this
happens, cf-agent may respawn cf-execd a few moments later, rendering
our killing ineffective. This is not completely foolproof, since both
cf-agent and cf-execd can spawn the other, so after killing one of
them, we need to check for the other. But it lowers the chance of it
happening significantly.

Also remove cf-hub killing, since there is no such process on AIX.