Make sure zlib is not a dependency of libxml for Solaris and HPUX.
On Solaris 9 there is some strange corruption (miscompilation?) when
libxml tries to use zlib, which leads to segfault. We don't need it so
turn it off completely.

HPUX is also changed at the same time, since it's relatively new, and
I want to reduce the build complexity. The others have built with it
a long time though, so not changing those.