Fix Solaris binaries not finding the libnet library.
The LD_LIBRARY_PATH points to Java libraries, and for some reason the
OpenLDAP build thinks it needs the libnet library from there, which is
of course completely wrong, since it is a Java library. Nevertheless
it tries to depend on it, which then causes problems when trying to
run the binary.
(cherry picked from commit 8f00e4f10fe387136b3c27fea5b86b831c963dad)