Redmine #6402: Consolidate nova-hub install scripts.
This change is basically the same as 9dbcdfebd9c, but for the
cfengine-nova-hub package. It consolidates the package scripts, and
takes out the most used platform specific parts and replaces them
with function calls. From now on only one script needs to be edited to
change the install scripts, but beware that it will be running on all
platforms, so use platform guards where you need to. See the mentioned
commit for more details.

It also adds some more introduction friendly documentation to the
README file about package scripts.