Backport commit 2fa5c1da5f49 from master.
Redmine #7045: Move cfengine3-web.service to core.

We don't need it there, but systemd makes it impossible to make a
conditional dependency on a service, so we need to keep the file in
core and put the conditional inside the actual service instead.

In master the service file is installed by automake, in 3.6 RHEL7 is a
separate build platform and must be handled by manual copying.