Change the command to create Design Centre's .gitignore
On looking at the installed file on my Debian test machine, I found it
started with the -e we'd passed to echo to get it to honour the '\n'
that appear in what we're echoing.  It turns out bash's echo outputs a
'\n' as such unless given -e; while dash's echo interprets \n as a
newline without a -e and includes any -e passed to it as part of its
output.  So no single echo invocation works for both.  Break up the
lines into separate invocations of echo, bung them in a sub-shell and
drop the sub-shell's accumulated output into .gitignore