Add build-scripts/bootstrap-tarballs and build-scripts/unpack-tarballs
The first is supposed to run from the "bootstrap" jenkins job and create
the core and masterfiles tarballs, and delete the git-checked-out
directories.

The second is supposed to run in the next stage in each of the
buildslaves, and unpack the tarballs and symlink into the core/ and
masterfiles/ directories.

This way we have a separate tarball from the "bootstrap" slave suitable
for releasing, and we test in all platforms that the tarball is valid.

(cherry picked from commit 8c08a2805b118e3e98e4e17201f5fe23f51ed5f5)