Redmine #7193: Fix an error in logic when deciding to run removal script
Upgrade is not detected correctly with bff and depot packages when
removing the package, because they use the existence of the binaries
to detect this. Obviously the binaries will exist when removing a
package, but this doesn't mean it's an upgrade. However, they don't
need this detection, because they don't run the removal scripts on
upgrade, so we can avoid this logic altogether.

Changelog: Fixed bug which would cause bff and depot packages not to
run package scripts on removal.