Redmine #6954: Prevent accumulation of unrelated AIX packages.
The directory is not cleaned, so each time we build a different
version we get a new package. This propogates all the way to the final
package directory, which is both confusing and potentially dangerous,
since someone may be tempted to download a package for a given version
that looks legitimate, but is in fact, an old/broken build.