Fix incorrect install script order on AIX during upgrade.
The preinstall script should be run before the package files are
removed, but on AIX they ran after. Solved by moving the preinstall
script to the pre_rm stage of the bff package. This required some
hacks since we cannot tell inside the script which one we're in, and
for installs we still need to run the normal preinstall stage.

Changelog: Fix daemons not restarting correctly on upgrade on AIX.
(cherry picked from commit 1bf19bdced0038955845ad9ca52194d0cc101760)