Redmine #6480: Enable libcurl for all platforms.
This is so it can be used by the url_get() policy function.

Note that we make a separate curl package for the hub. This is because
the hub will include the curl binary, but we don't want that for the
clients. It also introduces technical difficulties because on Windows
we get:

  /var/cfengine/bin/curl != /var/cfengine/bin/curl.exe