Fix certain Linux build targets bleeding over into other build slaves.
The problem is that a few select build targets are built even when
doing just "make dist". One example is the
tests/unit/init_script_test_helper, which is built because the test
shell script depends on it. This is no problem for the tarball,
because it won't be included, however when we afterwards copy the
workspace to the other build slaves, they will think that this target
was already built, and therefore it won't be rebuilt in the native
binary format.  Obviously a Solaris host won't be able to run a Linux
binary.
(cherry picked from commit 434c375873d9d8c6a97d04843f88890b217afecc)