ENT-2724: Fix crash in PHP cleanup function.
Both Curl and OpenSSL want to call this function, but it's only
allowed to call it once. This is a bug in PHP itself, so remove for
now. A memory leak in the cleanup is insignificant compared to a
crash.
(cherry picked from commit c6d186d35fc0ea1b1d23b1ef0221c3f83f6f533c)