Upgrade CFEngine dependencies.
Changelog: Upgrade CFEngine dependencies to the following versions:
Core:
- OpenSSL 1.0.2j
- LibYAML 0.1.7
- Curl 7.50.3
Enterprise:
- PostgreSQL 9.6.0
- Redis 3.2.4
- PHP 5.6.26
- Git 2.10.1

Also rearrange Postgres dependency packages. Get rid of the split
packages, which we don't use anyway, and gather everything in main
and devel packages, like is done for the other dependencies. This
allows us to use more catch-all entries in the package manifests,
which is a good thing, because the lists tend to change a lot between
minor versions.

Signed-off-by: Kristian Amlie <kristian.amlie@cfengine.com>