INF-422: Symlink libgnurx-0.dll to the dir where test can find it.
OpenLDAP tests for regular expressions during its ./configure phase.
To achieve that, it compiles and runs a test executable. On mingw, it
is a Windows *.exe file, which expects to find the *.dll file in one
of predefined places, and /var/cfengine/bin is not among them.

We can not use WINEDLLPATH env here, because it works only when
executing files like this: `wine app.exe`, not like this: `./app.exe`.
I've heard that it's a security feature:
https://forum.winehq.org/viewtopic.php?p=2270#p2270

Signed-off-by: Aleksei Shpakovskii <aleksei.shpakovskii@cfengine.com>
(cherry picked from commit 0202342e57431abe9fc6ea92ad4ba2b0e977c4d1)