ENT-3129: Run database migrations during post-install
This runs database migrations during post-install. Note, it always tries
to move forward (even though migrations support going backwards with
the "down" option). If a package is either upgraded or downgraded then
the migration runs. We don't explicitly support downgrading packages
currently so I believe this is OK even if not ideal.