update openssl from 1.0.2n to 1.1.0g
Also, install Perl from source when pre-existing version is not good enough.

Fixing linker error while building perl on RHEL4 and Debian 4.

The patch is a commit from upstream, that will be released in perl-5.26.2.

And include openssl.cnf so that on install cfengine-nova-hub can generate hub.cert if needed