ENT-3595: Add one more extra option for the AIX linker
For some reason, the AIX linker doesn't export the symbols
prefixed with underscore. In combination with how OpenSSL does
its magic with ASM the result is that OpenSSL fails to link.

We need to add one more explicit option to tell the linker to
export everything. Not sure if this is the best thing to do, but
it works.