Increase the level of parallelism to 2 * NUM_CPUS
All our operations are IO-bound so having twice the number of
jobs as CPUs is not a problem. The rare cases (if there are even
any) where this makes things slightly slower due to
process/context switching are well worth the extra speed.

Also make sure a temporary value for the MAKEFLAGS variable is
really temporary.