Base our customized postgresql.conf on up to date upstream config
Instead of storing our own customized config file which is
probably based on who-knows-how-old version of the upstream
default config file we should just store the changes that **we
know** we want to apply. This way we don't introduce confusion by
hiding config options added in new versions and we can be sure
that our modifications are still applicable (otherwise the patch
would fail to apply during build).