ENT-4089: Try to preserve postgresql.conf on upgrades
When upgrading we should try to continue using the same
postgresql.conf file if user did some modifications. However, if
it fails, we should use our recommended file (if the system has
enough RAM) or the defaults (if not).

In both cases we should keep both versions of the config and let
the user know they should review the config and update it
accordingly.