Upgrade: assume PostgreSQL is running on 3.7.x if running cf-hub
3.7.x has broken detection of the PostgreSQL status and
'postgres' is always reported as 'not running'. Thus when we
capture the pre-upgrade state of runnig processes, PostgreSQL is
missing there and as a consequence it is not started when the
upgrade is finished.

To workaround this, let's assume that PostgreSQL is running on
3.7.x if cf-hub is also running.

Ticket: ENT-4082
(cherry picked from commit 5f0f0d691b41c0035d6021cb240ad083c49ad313)