Try to avoid some race conditions when checking if PostgreSQL is running
The process can terminate while we move from checking if the port
is occupied and getting more info about the process. It's
impossible to get rid of all the race conditions there, but we
can at least avoid some of the cases by reusing results instead
of getting new inconsistent ones.

Also, let's check if we can get PostgreSQL's status before trying
to stop it. If we cannot get the status, it is probably
terminated already.

Ticket: ENT-4082