Get running services from lines matching exactly what we expect
When saving state we are extracting names of CFEngine services
that are running. The recognized sevices are the ones starting
with the 'cf-' prefix and the 'cfengine3.service'. Everything
else has to be ignored, otherwise the 's//' sed command fails and
we get full lines from the 'systemct list-units' output in the
state snapshot which is supposed to only list the running
services and nothing else.

Ticket: ENT-4082
(cherry picked from commit 3516c52761405f20af2b259ac2a44bd08624abec)