ENT-2653 Double default max_stack_depth in postgresql.conf
This change doubles the default max_stack_depth from 2MB to 4MB in
postgresql.conf.

According to the PostgreSQL documentation:

"The ideal setting for this parameter is the actual stack size limit enforced by
the kernel (as set by ulimit -s or local equivalent), less a safety margin of a
megabyte or so. The safety margin is needed because the stack depth is not
checked in every routine in the server, but only in key potentially-recursive
routines such as expression evaluation. The default setting is two
megabytes (2MB), which is conservatively small and unlikely to risk crashes."

https://www.postgresql.org/docs/10/static/runtime-config-resource.html#GUC-MAX-STACK-DEPTH

I found many references to setting this to 10MB but I found that some modern
systems are defaulting stack size to 8MB, so 4MB seems like a good increase, but
still safely conservative without implementing dynamic calculation.

Changelog: Title