Keep the .a files from OpenSSL on AIX
These are archives with shared objects which AIX needs for
runtime linking.

Unfortunately, rpm doesn't support using '%if' inside the
'%files' section so we need to do the decision in the '%install'
section, store the results in a helper file and then use the
helper file as part of the info for the '%files' section.

Ticket: ENT-4212
(cherry picked from commit 3a62be3fa984bd3a3ddd61509825de0a9d14582a)