Refactor how we check if PostgreSQL should be prepared for migration
Having the same check in two places is error prone to not
updating both places with changes. Exactly as it happened
recently in 7c515c832b.

Doing the check once and storing the result is better and will
also make it easier to change the check to something more complex
doing decisions on version differences and not hardcoded values.