Give PostgreSQL processes extra time to terminate
Some older versions of pg_ctl seem to leave a couple postgres
processes running when 'pg_ctl stop' finishes. In a second, they
are all gone. But if we are too quick checking if there is
something listening on the port 5432, we may stop upgrade process
for no good reason.

Changelog: None