Tell Perl to not use nexttoward() on AIX 7
This 128bit floating point API is missing on AIX, but the
function is declared in the header files. Perl sees the function
there and tries to use it, but at runtime, the function is not
provided by any of the standard libraries.

Ticket: ENT-4659
Changelog: None
(cherry picked from commit 8e8efac2a627fed2791b26540774be70ae46a048)