Fixed OOTB https redirection when hostname -f contains uppercase
The output of `hostname -f` was used as part of the ssl certificate filename
while the mustache template rendered by cf-agent uses sys.fqhost which is always
lowercase. This mis-alignment resulted in the rendered httpd.conf not
validating, therefore not updating the live configuration to redirect all http
traffic to https.

Changelog: Title
Ticket: ENT-4821