Install and load CFEngine Enterprise SELinux policy on RHEL 8
RHEL 8 has SELinux in the enforcing mode by default with
transitions from systemd being confined. So all our daemons are
running in confined domains as a confined user and thus we need
to ship a valid SELinux policy that allows them to do what we
know they need to do and that allows agent to do anything (by
running in a separate unconfined domain).