Fix key rotation on an HA setups
HA setup means that PostgreSQL is not handled by the init scripts
or systemd units. So to do the key rotation, PostgreSQL needs to
be started explicitly for the key rotation and then
stopped. Also, PostgreSQL is in the recovery (read-only) mode on
the passive hub and attempts to INSERT or UPDATE rows there only
result in noise.

Ticket: ENT-4695