ENT-4074 Better delete dirs and abort if it didn't succeed
Issue was that sometimes build aborted on one of following lines with
'ln: core exists' error. Assuming that it's because `rm` command left it behind,
let's help it by elevating `rm`'s priviledges, and printing more meaningful
error message if it fails, together with list of items if these dirs to aid
debugging this issue should it appear again.