Added creation of postgresql.log to nova hub package build
rpmbuild fails if a listed file is not present at build time. This change makes
sure the file exists in the build root. Also, having the postgresql.log listed
in files will make rpm understand that the CFEngine hub package actually owns
the file, which is a great improvement.

Ticket: ENT-7966
Changelog: None