Fixed behavioral comment, improved logging of issues encountered
It's kind of nice to know why what's happening. Reading the log prior to this
change, you see that something is copied and deleted and copied and maybe
deleted again for why? This simply prints out some messages in the event of
failure so that it's a bit easier to read the logs.

(cherry picked from commit eb13ca5b37e5ffc92ac695e6c92dbd0f6dc59571)