Check for existence of pg/data/postgresql.conf file, instead of data dir
Issue was that when /var/cfengine/state/pg/data is a mounted volume,
it can't be deleted. First, it makes `rm -rf` fail, and second, it
confuses migration logic which checks lack of this directory as a signal
to create new postgresql database there (including postgresql.conf
file). But when this directory is a mounted volume - `rm -rf` can delete
everything inside it, except the directory itself. Empty directory is
not a valid postgresql database, what confuses rest of postinstall script.

Note that we don't want to use plain is_upgrade here, because the
generate_new_postgres_conf must be executed in two cases:
* on a fresh install
* when a major upgrade of PostgreSQL requires DB migration - happens
  only during minor CFEngine upgrade (3.15 -> 3.18)

and is_upgrade is, well, _any_ CFEngine upgrade.

Ticket: ENT-8423