libcurl: disable ntlm
Ticket: ENT-8933

Originally it was done to remove some encryption method, but later that
encryption method was enabled back for php, so now ntlm can be enabled
on libcurl also. On the other side, this is some Windows-specific thing,
and our build of curl doesn't need it. So let's keep it disabled unless
someone objects.