Take 64-bit library on sparc and 32-bit library elsewhere
Note that this assumes that we build only 64-bit binaries on sparc,
and 32-bit binaries on other platforms. As of writing, this is correct
("other platforms" being only 32-bit x86 Intel).

Ticket: ENT-9013

Issue was that recently we switched to building 64-bit binaries on sparc
(apparently they can run both 32 and 64-bit), but we still were packaging
32-bit library, and thus all our binaries were crashing on start.