Added libexpat as explicit dependency
Issue was that until now it was required implicitly (invisibly).
For example, on Ubuntu 20, `ldd /var/cfengine/httpd/bin/httpd` shows,
among other things:

        libexpat.so.1 => not found

This seems to be an integral part of Apache httpd web server, and a
dependency which can't be disabled: Apache requires both apr and
apr-util[1], and apr-util requires libexpat[2].
[1]: https://github.com/apache/httpd/blob/trunk/INSTALL#L56
[2]: https://github.com/apache/apr-util/blob/1.6.1/configure.in#L169

Note that next release of apr-util will likely support libxml[3], a
library which we already build and bundle! So we could just wait until
it's released. But given that previous release of apr-util[4] was in
2017... This might be a bit of a long wait.
[3]: https://github.com/apache/apr-util/commit/50b8ac8146ef0060d6d5fe384127ca55b7380d3c
[4]: https://apr.apache.org/index.html

Also worth noting that it was in apr-util version 1.6.0 when they
unbundled[5] expat into a separate installable, so it is a CFEngine
3.12.0 regression[6].
[5]: https://downloads.apache.org/apr/CHANGES-APR-UTIL-1.6
[6]: https://github.com/cfengine/buildscripts/commit/3d9b06f94d2ad15ffd875cb316921667db1fb551

A ticket to build and package this library (a.k.a. "proper fix") is
ENT-9137. This is just a temporary fix so installation of a hub package
would fail _loudly_ (and user would understand what's missing), rather
then _silently_ somewhere in our postinstall script, leaving users
wondering and investigating.

Ticket: ENT-9105
(cherry picked from commit 203f543b0ec03bfa6353d8098fedd4e9254f6b87)