Added note about why SSLCompression is not explicitly off
When this was removed from buildscripts it also needed to be removed from the
template in masterfiles. We spent some time hunting down the issue before we
realized this was a change we made, not an upstream difference. Rather than
removing it from the config in masterfiles completely we stub it out with an
explanation.

Ticket: ENT-9600
Changelog: None