Make gzip a weak dependency
Investigation shows that the only place where it's _required_ was when
compressing man pages. But `man` can also work with uncompressed pages.

Other places which use gzip are Federated Reporting, watchdog script,
and self-upgrade on Solaris, - all these parts are not "essential".

So it was decided to downgrade gzip requirement to "Recommends".

Also worth noting that it's essential part of system on many
distibutions, and it's pretty hard to find a platform without gzip.

Ticket: ENT-9113