Revert "Added assert on EINVAL in mdb_mutex_failed()"
This reverts commit 43434600f675b352b1fcbd41ef3054df3270c397.

Unfortunately, in ENT-9835 we find that dependencies aren't aware if they
are involved in a RELEASE build and so should have NDEBUG defined.
So revert this change and later work on adding this capability so we can
have asserts in dependencies for CI/debug builds and not RELEASE builds.

(cherry picked from commit 8413a9a28021b214358cc2d60a10b13e797fc17a)