Refactored sftp cache ssh key usage into the docker container for full pull/push usage
Before I was pulling the sftp cache on the host outside the container because
I couldn't figure out how to push the sftp cache private key inside the container.

When I changed a dependency, zlib in this case, the build scripts wanted to push
the new packages up to the sftp cache which failed due to lack of private key in the container.

This commit takes either a github secret or mystiko password and sets it as an environment variable inside the container for the build to use.

Ticket: ENT-10419
Changelog: none
(cherry picked from commit b989b9d77d3c14b800c1ceb27c7e64237f9ee06d)