Modified hub package install scriptlet to use hostname -s and fail if that is longer than 64 characters
We create a self-signed certificate and the CN must be 64 characters or less so use hostname -s instead of hostname -f and fail if even the short name is longer than 64 characters.

This check is added to the preinstall scriptlet so that the package will not even be unpacked if hostname -s is longer than 64 characters long.

This check is only activated if there is no current cert present such as during an upgrade.

Ticket: CFE-4469
Changelog: title

libre