Fixed issue where rhel >8 packages would not have correct openssl dependency version
We build against systems with the latest available dependencies such as OpenSSL.
We use rpm -q --provides to determine the highest API present in OpenSSL and then use that as a Requires.
OPENSSL_VERSION is determined in build-scripts/package script.
This should ensure that when packages are installed with yum/dnf any required OpenSSL package upgrades will be performed or the installation will fail.

Ticket: ENT-12587
Changelog: title

libre