Adjusted mingw libxml2 build to exclude libiconv
In libxml2 2.13.6 and previous the configure script would allow no iconv to be found.
With 2.14.1 the configure script was changed to default to looking for iconv if no explicit --without-iconv was specified.

We have been building libxml2 on mingw platform without iconv "always" as far as I can tell so this is not a functional change, only a build-time change.

Ticket: ENT-12744
Changelog: none