update-deps.yml: fetch update-deps.py from master
The workflow for updating dependencies keeps failing on the weekly
schedule because it checks out the `3.21.x` branch that does not have
the update-deps.py script. Instead of coping this script to the older
branch, I thought it would make more sense to fetch it from the master
branch. This way we only need to maintain one copy of it.

Ticket: CFE-4530
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>