version: document and simplify
How can it be so simple? Well, since detect-version.sh replaced
detect-version.py, the CFVERSION will always contain exactly what is
needed. Furthermore, the CODE_COVERAGE built type is not being used. And
if we decide to use it, we don't need code_coverage to be appended to
the version number. On top of that, the "new" detect-version.sh will
never add dash characters to the version number. Hence, we don't need to
replace them with tilde characters for the packaging tools to work.

Ticket: ENT-12600
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>