compile-option: Remove auto detect role code
This auto detect role code is weird. It tries to detect whether to build
a hub or agent package based on the platform it's being built for. In
many places it even decides to build hub packages for platforms where
it's not supported. I wonder if life will just go on if I remove it.

Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>