Fixed bug in grep_c function for exotics
The grep_c function is supposed to do a platform independent grep with
context. However, all the `-A`, `-B`, & `-C` arguments are not available
on exotics. Thus, the function would perform a grep without context on
these platforms. The problem is that I forgot to remove the `NUM`
argument when grepping without context.

Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>