build-environment-check: deleted script
The list of wanted/unwanted packages is outdated. It only handles redhat
and debian based systems. Nowadays we use the
cfengine-build-host-setup.cf as the single source of truth. It does
exactly what we want, it promises that the build host is setup
correctly. Hence, it's time we let this script go.

Ticket: ENT-12598
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>