Increased general timeout for Mission Portal httpd.conf to allow for slower hardware and longer agent runs
The original issue was that trigger agent run and collect reports from MP UI was timing out due to long agent runs.

The fix was in place in masterfiles at https://github.com/cfengine/masterfiles/pull/3051 but we neglected to add it to the distribution httpd.conf here in buildscripts.

Ticket: ENT-13291
Changelog: title